use composer commands in travis (single point of maintaince
move LDAP tests in a separate tree
split tests for Telemetry, and skip test when plugins are installed
improve test when plugins are installed
Rename api to web in tests (no longer concerns only API)
Split tests for LDAP
