+++ .travis.yml
@@ -27,10 +27,11 @@ before_script:
   - phpenv config-add tests/enable-apcu.ini
 script:
   - mysql -u root -e 'select version();'
-  - vendor/bin/atoum --debug --configurations tests/telemetry.php --bootstrap-file tests/bootstrap.php --no-code-coverage --max-children-number 1 -d tests/database
-  - vendor/bin/atoum --debug --configurations tests/telemetry.php --bootstrap-file tests/bootstrap.php --no-code-coverage --max-children-number 1 -d tests/units
-  - vendor/bin/atoum --debug --configurations tests/telemetry.php --bootstrap-file tests/bootstrap.php --no-code-coverage --max-children-number 1 -d tests/api
-  - if [[ "$CS" == "true" ]]; then vendor/bin/phpcs -p --standard=vendor/glpi-project/coding-standard/GlpiStandard/ --ignore=/vendor/,/plugins/,/files/,/lib/,/config/,/tests/config ./; fi
+  - composer testdb
+  - composer test
+  - composer testweb
+  - composer testldap
+  - if [[ "$CS" == "true" ]]; then composer cs; fi
 
 #note: default maria version is 5.5 for all main php versions list exept nightly
 matrix:
