+++ tests/LDAP/README.md
@@ -1,0 +1,1 @@
+Unit tests that require a runing LDAP server
