+++ tests/web/README.md
@@ -1,0 +1,1 @@
+Unit tests that require a running webserver
